+++ src/fields/formfields/Name.php
@@ -4,8 +4,8 @@ namespace barrelstrength\sproutforms\fields\formfields;
 
 use barrelstrength\sproutbasefields\SproutBaseFields;
 use barrelstrength\sproutforms\rules\conditions\ContainsCondition;
-use barrelstrength\sproutforms\rules\conditions\DoesNotContainsCondition;
-use barrelstrength\sproutforms\rules\conditions\DoesNotEndsWithCondition;
+use barrelstrength\sproutforms\rules\conditions\DoesNotContainCondition;
+use barrelstrength\sproutforms\rules\conditions\DoesNotEndWithCondition;
 use barrelstrength\sproutforms\rules\conditions\DoesNotStartWithCondition;
 use barrelstrength\sproutforms\rules\conditions\EndsWithCondition;
 use barrelstrength\sproutforms\rules\conditions\IsCondition;
@@ -302,11 +302,11 @@ class Name extends FormField implements PreviewableFieldInterface
             new IsCondition(),
             new IsNotCondition(),
             new ContainsCondition(),
-            new DoesNotContainsCondition(),
+            new DoesNotContainCondition(),
             new StartsWithCondition(),
             new DoesNotStartWithCondition(),
             new EndsWithCondition(),
-            new DoesNotEndsWithCondition()
+            new DoesNotEndWithCondition()
         ];
     }
 }